+++ frontend/src/pages/PromotionPage/PromotionDetailPage.tsx
@@ -1,6 +1,6 @@
+import { useParams } from 'react-router-dom';
 import { PAGE_VIEW } from '@/constants/eventName';
 import useTrackPageView from '@/hooks/Mixpanel/useTrackPageView';
-import { useParams } from 'react-router-dom';
 
 const PromotionDetailPage = () => {
   useTrackPageView(PAGE_VIEW.PROMOTION_DETAIL_PAGE);
