+++ frontend/src/pages/PromotionPage/components/PromotionGrid/PromotionGrid.styles.ts
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import styled from 'styled-components';
 
 export const Grid = styled.div`
   display: grid;
